Move WithI18n container to shared
