+++ app/src/marketplace/containers/LocaleSetter/index.jsx
@@ -5,7 +5,7 @@ import qs from 'query-string'
 import { applyLocale, getDefaultLocale } from '$shared/utils/locale'
 import { withRouter } from 'react-router-dom'
 
-import withI18n, { type I18nProps } from '../WithI18n'
+import withI18n, { type I18nProps } from '$shared/containers/WithI18n'
 import i18n from '../../i18n'
 import store from '../../../store'
