+++ app/src/shared/components/Layout/Footer.jsx
@@ -7,7 +7,7 @@ import { connect } from 'react-redux'
 import { withRouter } from 'react-router-dom'
 import { push } from 'connected-react-router'
 
-import withI18n, { type I18nProps } from '$mp/containers/WithI18n'
+import withI18n, { type I18nProps } from '$shared/containers/WithI18n'
 import { formatPath } from '$shared/utils/url'
 import routes from '$routes'
 import links from '$shared/../links'
